csv record attempt 1
Help me create a button in this code start recording the blend shapes values into a csv file, that will be exported after pressing to stop the recording. After pressing the recording button, a counter will appear, after pressing stop, the counter will be set to zero, and a pop up will appear to choose the location and name for the csv file created. The csv file will have the blend shape values for each frame.